feat: unify Python API with MCP tools for all modules
Extends the pattern from PR #176 (LLM tools) to all remaining MCP modules.
Core functions now live in shared.py with identical interfaces to MCP tools
but without Field() decorators, making them directly importable for Python use.
MCP tools become thin wrappers that only add Field() metadata.

Modules unified:
- google_calendar: read, create, update, delete
- google_maps: get_directions, server_info
- email: read, update, send, sync
- screenshot: grab
- microsoft.word: read, edit, create, render
- microsoft.excel: read, edit

Also adds deprecation warnings to arxiv, code2prompt, py2nb.
