+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "mcp-handley-lab"
-version = "0.24.0"
+version = "0.24.1"
 description = "MCP Handley Lab - A comprehensive MCP toolkit for research productivity and lab management"
 readme = "README.md"
 requires-python = ">=3.10"
@@ -90,7 +90,6 @@ mcp-screenshot = "mcp_handley_lab.screenshot.tool:mcp.run"
 mcp-word = "mcp_handley_lab.microsoft.word.tool:mcp.run"
 mcp-excel = "mcp_handley_lab.microsoft.excel.tool:mcp.run"
 mcp-mathematica = "mcp_handley_lab.mathematica.cli:main"
-mcp-search = "mcp_handley_lab.search.tool:mcp.run"
 mcp-handley-lab = "mcp_handley_lab.__main__:main"
 mcp-cli = "mcp_handley_lab.cli.main:cli"
