+++ src/mcp_handley_lab/arxiv/__init__.py
@@ -1,1 +1,14 @@
-"""ArXiv source code retrieval tool."""
+"""ArXiv source code retrieval tool.
+
+.. deprecated::
+    This module is deprecated. Use the /arxiv skill instead.
+"""
+
+import warnings
+
+warnings.warn(
+    "mcp_handley_lab.arxiv is deprecated and will be removed. "
+    "Use the /arxiv skill instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
