+++ src/mcp_handley_lab/code2prompt/__init__.py
@@ -1,0 +1,14 @@
+"""Code2prompt codebase summarization tool.
+
+.. deprecated::
+    This module is deprecated. Use the /code2prompt skill instead.
+"""
+
+import warnings
+
+warnings.warn(
+    "mcp_handley_lab.code2prompt is deprecated and will be removed. "
+    "Use the /code2prompt skill instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
