+++ src/mcp_handley_lab/py2nb/__init__.py
@@ -1,1 +1,14 @@
-"""py2nb conversion tool for bidirectional Python script ↔ Jupyter notebook conversion."""
+"""py2nb conversion tool for bidirectional Python script ↔ Jupyter notebook conversion.
+
+.. deprecated::
+    This module is deprecated. Use py2nb directly instead.
+"""
+
+import warnings
+
+warnings.warn(
+    "mcp_handley_lab.py2nb is deprecated and will be removed. "
+    "Use py2nb directly instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
